pyproject.toml
[build-system]
# Hatchling = build backend moderne, simple, rapide
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

requires-python = ">=3.12"

[project]
name = "pyphpbb_sl"
version = "0.1.0"
authors = [{name = "Sergeileduc", email = "sergei.leduc@gmail.com"},]
maintainers = [{name = "Sergeileduc"}]
description = "Interract with phpbb forum."
license = {file = "LICENSE"}
requires-python = ">=3.12"
keywords = ["forum", "phpbb"]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
]

dependencies = [
    "httpx",
    "selectolax",
]

# ---------------------------------------------------------------------------

[dependency-groups]

test = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=5.0.0",
    "python-dotenv",
]

dev = [
    "ruff>=0.14.10",
    "mypy>=1.19.1",
    "coverage>=7.13.1",
    "invoke>=2.2.1",
]

full = [
    { include-group = "test" },
    { include-group = "dev" },
]


# ---------------------------------------------------------------------------

[project.urls]
"Homepage" = "https://github.com/Sergeileduc/pyphpbb-sl"
"Source" = "https://github.com/Sergeileduc/pyphpbb-sl"
"Bug Tracker" = "https://github.com/Sergeileduc/pyphpbb-sl/issues"
"Documentation" = "https://dpytest.readthedocs.io/"

[tool.setuptools]
packages = ["pyphpbb_sl"]

[tool.setuptools.dynamic]
readme = {file = ["README.md", "HISTORY.md"], content-type = "text/markdown"}
version = {attr = "pyphpbb_sl.__version__"}


##### 🛠 DEV TOOLS #####
[tool.ruff]
# Linter + formatter + import sorter
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

# ---------------------------------------------------------------------------

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_unused_ignores = true
warn_return_any = true
warn_redundant_casts = true
warn_unreachable = true
disallow_untyped_defs = false
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
ignore_missing_imports = true
plugins = []
files = ["."]
exclude = '''
(
    ^\.?venv.*$        # .venv, venv, venv11, .venv11, etc.
  | ^env.*$            # env, env3.12, etc.
  | ^tests?$           # test, tests
  | ^tests/.*$         # contenu du dossier tests
)
'''

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore:.*U.*:DeprecationWarning",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.coverage.run]
branch = true
source = ["python_tools_sl"]

[tool.coverage.report]
show_missing = true
skip_covered = true

##### 🛠 BUMP2VERSION #####

[tool.bumpversion]
current_version = "1.0.0"
commit = true
tag = true

[[tool.bumpversion.files]]
filename = "pyphpbb_sl/__init__.py"
search = "__version__ = \"{current_version}\""
replace = "__version__ = \"{new_version}\""
